refactor(controller): split dj-agent.ts's supporting parts into dj-agent/
dj-agent.ts 1282 -> 783. The two runs it exists for (runTrackEvent,
runRequest) and the fallback paths around them stay; what they're built
from moves out:

  runs.ts      DJ-mode mini-runs (a short arc of picks heading somewhere)
  schemas.ts   the pick/request output schemas and system prompts
  breaker.ts   the circuit breaker that drops to the pool picker
  agents.ts    the two tool-loop agent definitions
  enqueue.ts   turning a chosen song into a queued track

Worth flagging: scripts/llm-bench imports pickSystem/pickSchema/
pickerAgent (and the request trio) straight from dj-agent.js, and
scripts/ isn't in tsconfig's include — so a missing re-export there would
not have shown up as a type error. dj-agent.ts re-exports the full
original surface and a runtime import check confirms all twelve names
resolve.

No behaviour change. Controller lint clean, all 52 test files pass.
